+++ genprotimg/src/include/pv_hdr_def.h
@@ -25,6 +25,9 @@
 #define PV_VERSION_1	0x00000100U
 
 /* Plaintext control flags */
+#define PV_PCF_PCKM_ECC		(1ULL << 5)  /* PCKMO encrypt-ECC-key functions allowed */
+#define PV_PCF_PCKMO_AES	(1ULL << 6)  /* PCKMO encrypt-AES-key functions allowed */
+#define PV_PCF_PCKMO_DEA_TDEA	(1ULL << 7)  /* PCKMO encrypt-DEA/TDEA-key functions allowed */
 #define PV_PCF_NO_DECRYPTION	(1ULL << 28) /* prevent Ultravisor decryption during unpack operation */
 
 /* maxima for the PV version 1 */